Make FontDefinitions itself the configured FontProvider
Drop the `ConfiguredFonts` wrapper: `FontDefinitions` implements
`FontProvider` directly, and `FontsImpl` holds it in an `Arc` that it
shares with the provider list instead of cloning the definitions.
